+++ pybind/minkowski.cu
@@ -25,6 +25,7 @@
  * of the code.
  */
 #include "extern.hpp"
+#include "ME_build_info.h"
 
 #include <string>
 
@@ -34,6 +35,7 @@
 #include <pybind11/stl.h>
 
 PYBIND11_MODULE(TORCH_EXTENSION_NAME, m) {
+  m.attr("_build_info") = ME_BUILD_INFO;
   // Constant function
   m.def("is_cuda_available", &is_cuda_available);
   m.def("cuda_version", &cuda_version);